The Complete Collection
For an amazing 3 or so months development of this python code has been
rather fruitful, but not its starting to become a bit of a mess. Time to
implement version control! Na well there's still much to do in the long
road of analysing this data including knowing what the cloud type is!
You keep forgetting about this little add-on didn't you ;)